app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
celery
redis
